+++ test/test-controller.js
@@ -1,5 +1,5 @@
 const {Choices}            = require('../src/lib/calibration.js')
-const {Controller, Pages}  = require('../src/lib/controller.js');
+const Controller           = require('../src/lib/controller.js');
 const Model                = require('./helpers/model.js');
 const View                 = require('./helpers/view.js');
 global.winston             = require('./helpers/winston.js');
@@ -12,21 +12,21 @@ test('Controller without calibration basically works', async t => {
   let c = new Controller(model, view);
   await c.restart();
 
-  t.is(view.wizard.index.page, Pages.SETTINGS);
+  t.true(view.wizard.index.isSettingsPage());
   await c.next();
-  t.is(view.wizard.index.page, Pages.CHOOSE_CAL);
+  t.true(view.wizard.index.isChooseCalPage());
   view.wizard.calChoice = Choices.NONE
   await c.next();
-  t.is(view.wizard.index.page, Pages.MEASURE);
+  t.true(view.wizard.index.isMeasurementPage());
   const procedure = await model.getProcedure();
   const NUM_STEPS = procedure.steps.length;
   for (let i = 0; i < NUM_STEPS; i++) {
-    t.is(view.wizard.index.page, Pages.MEASURE);
+    t.true(view.wizard.index.isMeasurementPage());
     t.is(view.wizard.index.step, i);
     await c.next();
   }
   await c.next();
-  t.is(view.wizard.index.page, Pages.MEASURE);
+  t.true(view.wizard.index.isMeasurementPage());
   t.is(view.wizard.index.step, NUM_STEPS-1);
 });
 
@@ -38,27 +38,27 @@ test('Controller with saved cal basically works', async t => {
 
   let c = new Controller(model, view);
   await c.restart();
-  t.is(view.wizard.index.page, Pages.SETTINGS);
+  t.true(view.wizard.index.isSettingsPage());
   view.vnaAddress        = '127.0.0.1';
   view.matrixAddress     = '1.2.3.4';
   view.procedureFilename = './test/fixtures/procedures/procedure.yaml';
   await c.next();
-  t.is(view.wizard.index.page, Pages.CHOOSE_CAL);
+  t.true(view.wizard.index.isChooseCalPage());
   view.wizard.calChoice = Choices.EXISTING
   view.wizard.calGroup  = 'cal group 2';
   await c.next();
   t.is(model.calChoice, Choices.EXISTING);
   t.is(model.calGroup, 'cal group 2');
-  t.is(view.wizard.index.page, Pages.MEASURE);
+  t.true(view.wizard.index.isMeasurementPage());
   const procedure = await model.getProcedure();
   const NUM_STEPS = procedure.steps.length;
   for (let i = 0; i < NUM_STEPS; i++) {
-    t.is(view.wizard.index.page, Pages.MEASURE);
+    t.true(view.wizard.index.isMeasurementPage());
     t.is(view.wizard.index.step, i);
     await c.next();
   }
   await c.next();
-  t.is(view.wizard.index.page, Pages.MEASURE);
+  t.true(view.wizard.index.isMeasurementPage());
   t.is(view.wizard.index.step, NUM_STEPS-1);
 });
 
@@ -69,38 +69,36 @@ test('Controller performing calibration basically works', async t => {
 
   let c = new Controller(model, view);
   await c.restart();
-  t.is(view.wizard.index.page, Pages.SETTINGS);
+  t.true(view.wizard.index.isSettingsPage());
   view.vnaAddress = '127.0.0.1';
   view.matrixAddress = '1.2.3.4';
   view.procedureFilename = './test/fixtures/procedures/procedure.yaml';
   await c.next();
-  t.is(view.wizard.index.page, Pages.CHOOSE_CAL);
+  t.true(view.wizard.index.isChooseCalPage());
   view.wizard.calChoice = Choices.CALIBRATE;
   await c.next();
   t.is(model.calChoice, Choices.CALIBRATE);
-  t.is(view.wizard.index.page, Pages.CALIBRATE);
+  t.true(view.wizard.index.isCalibrationPage());
   t.is(view.wizard.index.step, 0);
   await c.next();
-  console.log(view.alert);
-  t.is(view.wizard.index.page, Pages.CALIBRATE);
+  t.true(view.wizard.index.isCalibrationPage());
   t.is(view.wizard.index.step, 1);
   await c.next();
-  t.is(view.wizard.index.page, Pages.CALIBRATE);
+  t.true(view.wizard.index.isCalibrationPage());
   t.is(view.wizard.index.step, 2);
   await c.next(); // Perform and apply
-  for (let i of c.history) {
-    t.not(i.page, Pages.CALIBRATE);
-  }
+  t.true(view.wizard.index.isMeasurementPage());
+  t.is(view.wizard.index.step, 0);
 
   const procedure = await model.getProcedure();
   const NUM_STEPS = procedure.steps.length;
   for (let i = 0; i < NUM_STEPS; i++) {
-    t.is(view.wizard.index.page, Pages.MEASURE);
+    t.true(view.wizard.index.isMeasurementPage());
     t.is(view.wizard.index.step, i);
     await c.next();
   }
   await c.next();
-  t.is(view.wizard.index.page, Pages.MEASURE);
+  t.true(view.wizard.index.isMeasurementPage());
   t.is(view.wizard.index.step, NUM_STEPS-1);
 });
 
@@ -111,7 +109,7 @@ test('parameters: Non-existent cal group cleared', async t => {
 
   model.calGroup        = 'does not exist';
   model.state.calGroups = ['cal group 1', 'cal group 2'];
-  c.index.page          = Pages.CHOOSE_CAL;
+  c.index.setChooseCalPage();
   let params            = await c.parameters();
   t.is(params.calGroup, 'cal group 1');
 });
@@ -123,7 +121,7 @@ test('parameters: No cal groups clears cal group', async t => {
 
   model.calGroup        = 'does not exist';
   model.state.calGroups = [];
-  c.index.page          = Pages.CHOOSE_CAL;
+  c.index.setChooseCalPage();
   let params            = await c.parameters();
   t.is(params.calGroup, '');
 });