+++ CHANGELOG.md
@@ -4,7 +4,7 @@ R&S Calibrate Switch Matrix Change Log
 2017-09-12 Version 0.8.4
 ------------------------
 
-* Non-existent cal group cleared for choose cal group
+* Existing cal group bug(s) fixed. User choice persists despite back/next.
 * Procedure dialog opens `yaml` files only, starts in documented default location
 
 2017-08-06 Version 0.8.3