+++ CHANGELOG.md
@@ -6,6 +6,7 @@ R&S Calibrate Switch Matrix Change Log
 
 * Existing cal group bug(s) fixed. User choice persists despite back/next.
 * Procedure dialog opens `yaml` files only, starts in documented default location
+* Alerts are persistent and dismissible
 
 2017-08-06 Version 0.8.3
 ------------------------